requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.12.0
fredapi>=0.5.0
yfinance>=0.2.0
pyyaml>=6.0
joblib>=1.3.0
